requirements.txt
asgiref==3.8.1
dbus-python==1.3.2
Django==5.0.6
django-filter==24.2
fontawesomefree==6.5.1
PyGObject==3.48.2
python-escpos==3.1
pyusb==1.2.1
sqlparse==0.5.0
